+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Resolves:
     - &#35;311 Add field for 'requires registry rebuild'
     - &#35;312 Add field for 'requires cache clear'
 - &#35;330 During launch, drupal caches are cleared a lot
+- &#35;412 Return something like `Command [xx:xx:xx] 2 of 16 - Success` in Slack
 
 ## v2.0.3